+++ databases/alembic/versions/2976b41930ad_change_lcd_16x4_to_20x4.py
@@ -6,6 +6,7 @@ Create Date: 2019-02-17 18:25:31.478802
 
 """
 from alembic import op
+import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
@@ -32,6 +33,61 @@ def upgrade():
         '''
     )
 
+    with op.batch_alter_table("output") as batch_op:
+        batch_op.add_column(sa.Column('state_at_startup', sa.Boolean))
+        batch_op.add_column(sa.Column('state_at_shutdown', sa.Boolean))
+        batch_op.add_column(sa.Column('on_state', sa.Boolean))
+
+    op.execute(
+        '''
+        UPDATE output
+        SET on_state=NULL
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET on_state=0
+        WHERE trigger=0
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET on_state=1
+        WHERE trigger=1
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET state_at_startup=1
+        WHERE on_at_start=1
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET state_at_startup=0
+        WHERE on_at_start=0
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET state_at_shutdown=0
+        '''
+    )
+
+    with op.batch_alter_table("output") as batch_op:
+        batch_op.drop_column('on_at_start')
+        batch_op.drop_column('trigger')
+
 
 def downgrade():
     op.execute(
@@ -49,3 +105,44 @@ def downgrade():
         WHERE lcd_type='20x4_generic'
         '''
     )
+
+    with op.batch_alter_table("output") as batch_op:
+        batch_op.add_column(sa.Column('on_at_start', sa.Boolean))
+        batch_op.add_column(sa.Column('trigger', sa.Boolean))
+
+    op.execute(
+        '''
+        UPDATE output
+        SET trigger=0
+        WHERE on_state=0
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET trigger=1
+        WHERE on_state=1
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET on_at_start=1
+        WHERE state_at_startup=1
+        '''
+    )
+
+    op.execute(
+        '''
+        UPDATE output
+        SET on_at_start=0
+        WHERE state_at_startup=0
+        '''
+    )
+
+    with op.batch_alter_table("output") as batch_op:
+        batch_op.drop_column('state_at_startup')
+        batch_op.drop_column('state_at_shutdown')
+        batch_op.drop_column('on_state')